+++ src/main/scala/sdql/backend/codegen/package.scala
@@ -30,7 +30,7 @@ package object codegen {
       s"std::tuple$template"
     case BoolType                 => "bool"
     case RealType                 => "double"
-    case IntType | DateType       => "long"
+    case IntType | DateType       => "int"
     case StringType(None)         => "std::string"
     case StringType(Some(maxLen)) => s"VarChar<$maxLen>"
     case tpe                      => raise(s"unimplemented type: $tpe")
